[FLINK-4660] Allow CoProcessFunction on non-keyed ConnectedStreams
Introduce new CoProcessOperator for this. Rename the pre-existing
CoProcessOperator to KeyedCoProcessOperator.
