+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/datastream/ConnectedStreams.java
@@ -29,9 +29,10 @@ import org.apache.flink.streaming.api.functions.co.CoFlatMapFunction;
 import org.apache.flink.streaming.api.functions.co.CoMapFunction;
 import org.apache.flink.streaming.api.functions.co.CoProcessFunction;
 import org.apache.flink.streaming.api.operators.TwoInputStreamOperator;
+import org.apache.flink.streaming.api.operators.co.CoProcessOperator;
 import org.apache.flink.streaming.api.operators.co.CoStreamFlatMap;
 import org.apache.flink.streaming.api.operators.co.CoStreamMap;
-import org.apache.flink.streaming.api.operators.co.CoProcessOperator;
+import org.apache.flink.streaming.api.operators.co.KeyedCoProcessOperator;
 import org.apache.flink.streaming.api.transformations.TwoInputTransformation;
 
 import static java.util.Objects.requireNonNull;
@@ -281,13 +282,13 @@ public class ConnectedStreams<IN1, IN2> {
 			CoProcessFunction<IN1, IN2, R> coProcessFunction,
 			TypeInformation<R> outputType) {
 
-		if (!(inputStream1 instanceof KeyedStream) || !(inputStream2 instanceof KeyedStream)) {
-			throw new UnsupportedOperationException("A CoProcessFunction can only be applied" +
-					"when both input streams are keyed.");
-		}
+		TwoInputStreamOperator<IN1, IN2, R> operator;
 
-		CoProcessOperator<Object, IN1, IN2, R> operator = new CoProcessOperator<>(
-				inputStream1.clean(coProcessFunction));
+		if ((inputStream1 instanceof KeyedStream) && (inputStream2 instanceof KeyedStream)) {
+			operator = new KeyedCoProcessOperator<>(inputStream1.clean(coProcessFunction));
+		} else {
+			operator = new CoProcessOperator<>(inputStream1.clean(coProcessFunction));
+		}
 
 		return transform("Co-Process", outputType, operator);
 	}
